# docs/swagger.yaml
definitions:
  handler.CodeSnippetForm:
    properties:
      is_archived:
        type: boolean
      is_draft:
        type: boolean
      is_private:
        type: boolean
      program_language_id:
        type: string
      text:
        type: string
    type: object
  handler.ProgramLanguageForm:
    properties:
      name:
        type: string
    type: object
  handler.User:
    properties:
      username:
        example: johndoe
        type: string
    type: object
  model.CodeSnippet:
    description: CodeSnippet is the model representing a code snippet in the system.
    properties:
      codeSnippetID:
        type: string
      codeSnippetRatings:
        items:
          $ref: '#/definitions/model.CodeSnippetRating'
        type: array
      createdAt:
        type: string
      isArchived:
        type: boolean
      isDraft:
        type: boolean
      isPrivate:
        type: boolean
      programLanguage:
        $ref: '#/definitions/model.ProgramLanguage'
      programLanguageID:
        type: string
      reviewComments:
        items:
          $ref: '#/definitions/model.ReviewComment'
        type: array
      text:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userID:
        type: string
    type: object
  model.CodeSnippetRating:
    description: CodeSnippetRating is the model representing a rating for a code snippet.
    properties:
      codeSnippet:
        $ref: '#/definitions/model.CodeSnippet'
      codeSnippetID:
        type: string
      codeSnippetRatingID:
        type: string
      createdAt:
        type: string
      rating:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userID:
        type: string
    type: object
  model.ProgramLanguage:
    description: ProgramLanguage is the model representing a programming language
      in the system.
    properties:
      name:
        type: string
      programLanguageID:
        type: string
    type: object
  model.ReviewComment:
    description: ReviewComment is the model representing a comment on a code snippet.
    properties:
      codeSnippet:
        $ref: '#/definitions/model.CodeSnippet'
      codeSnippetID:
        type: string
      commentID:
        type: string
      createdAt:
        type: string
      line:
        type: integer
      replyCommentID:
        type: string
      text:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userID:
        type: string
    type: object
  model.User:
    description: User is the model representing a user in the system.
    properties:
      createdAt:
        type: string
      isActive:
        type: boolean
      isAnonymous:
        type: boolean
      updatedAt:
        type: string
      userID:
        type: string
      username:
        example: johndoe
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/code_snippet/:
    get:
      consumes:
      - application/json
      description: Get all code snippets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CodeSnippet'
            type: array
      summary: Get all code snippets
      tags:
      - Code Snippets
    post:
      consumes:
      - application/json
      description: Create a code snippet
      parameters:
      - description: Code Snippet information to create
        in: body
        name: code_snippet
        required: true
        schema:
          $ref: '#/definitions/handler.CodeSnippetForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CodeSnippet'
      summary: Create a code snippet
      tags:
      - Code Snippets
  /api/v1/code_snippet/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a code snippet by ID
      parameters:
      - description: Code Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CodeSnippet'
      summary: Delete a code snippet
      tags:
      - Code Snippets
    get:
      consumes:
      - application/json
      description: Get a single code snippet by ID
      parameters:
      - description: Code Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CodeSnippet'
      summary: Get a single code snippet
      tags:
      - Code Snippets
  /api/v1/program_language/:
    get:
      consumes:
      - application/json
      description: Get all program languages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ProgramLanguage'
            type: array
      summary: Get all program languages
      tags:
      - Program Languages
    post:
      consumes:
      - application/json
      description: Create a program language
      parameters:
      - description: Name of the program language
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/handler.ProgramLanguageForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ProgramLanguage'
      summary: Create a program language
      tags:
      - Program Languages
  /api/v1/user/:
    get:
      consumes:
      - application/json
      description: get details of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create a new user with the provided information
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      summary: Create a new user
      tags:
      - users
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User deleted
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: get details of user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Get single user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update user's information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User information to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Update an existing user
      tags:
      - users
swagger: "2.0"